requirements.txt
numpy>=1.18
pyserial>=3.4
